requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
coverage==4.4.1
diary==0.1.4
Django==1.11.2
django-nose==1.4.5
django-rest-swagger==2.1.2
djangorestframework==3.6.4
httpie==0.9.9
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
nose==1.3.7
olefile==0.44
openapi-codec==1.3.2
Pillow==4.2.1
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
selenium==3.4.3
simplejson==3.11.1
uritemplate==3.0.0
urllib3==1.22
